requirements.txt
pandas
sqlite3
python-dotenv
langchain-google-genai
langchain-core